refactor: mutable elements iterator helper
