+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/MultiSetImplTest.kt
@@ -2,11 +2,9 @@ package xyz.lbres.kotlinutils.set.multiset.impl
 
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
 import xyz.lbres.kotlinutils.list.IntList
-import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMutableMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.testutils.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.Test
-import kotlin.test.assertEquals
 
 class MultiSetImplTest {
     private fun <T> createSet(): (Collection<T>) -> MultiSetImpl<T> = { MultiSetImpl(it) }
@@ -73,29 +71,8 @@ class MultiSetImplTest {
 
     @Test
     fun testIterator() {
-        runMultiSetIteratorTests({ MultiSetImpl(it) }, { MultiSetImpl(it) })
-
-        // mutable elements
-        val mutableList1 = mutableListOf(1, 2, 3)
-        val mutableList2 = mutableListOf(0, 5, 7)
-        val listSet: MultiSet<IntList> = multiSetOf(mutableList1, mutableList2)
-
-        var listIter = listSet.iterator()
-        var listExpected = listOf(listOf(1, 2, 3), listOf(0, 5, 7))
-        var listValues: MutableList<IntList> = mutableListOf()
-        while (listIter.hasNext()) {
-            listValues.add(listIter.next())
-        }
-        assertEquals(listExpected, listValues)
-
-        mutableList2.clear()
-        listIter = listSet.iterator()
-        listExpected = listOf(listOf(1, 2, 3), emptyList())
-        listValues = mutableListOf()
-        while (listIter.hasNext()) {
-            listValues.add(listIter.next())
-        }
-        assertEquals(listExpected, listValues)
+        runMultiSetIteratorTests(createSet(), createSet())
+        runMultiSetMutableElementsIteratorTests(createSet())
     }
 
     @Test
