+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/impl/MutableMultiSetImplTest.kt
@@ -2,13 +2,11 @@ package xyz.lbres.kotlinutils.set.multiset.impl
 
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
 import xyz.lbres.kotlinutils.list.IntList
-import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.testutils.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.set.multiset.testutils.runMultiSetMutableElementContainsTests
 import xyz.lbres.kotlinutils.set.multiset.testutils.runMultiSetMutableIteratorTests
 import kotlin.test.Test
-import kotlin.test.assertEquals
 
 class MutableMultiSetImplTest {
     private fun <T> createSet(): (Collection<T>) -> MutableMultiSetImpl<T> = { MutableMultiSetImpl(it) }
@@ -73,6 +71,7 @@ class MutableMultiSetImplTest {
     }
 
     @Test fun testIsEmpty() = runMultiSetMutableIsEmptyTests(createSet(), createSet())
+
     @Test
     fun testGetCountOf() {
         runMultiSetMutableGetCountOfTests(createSet(), createSet())
@@ -82,28 +81,7 @@ class MutableMultiSetImplTest {
     @Test
     fun testIterator() {
         runMultiSetMutableIteratorTests(createSet(), createSet())
-
-        // mutable elements
-        val mutableList1 = mutableListOf(1, 2, 3)
-        val mutableList2 = mutableListOf(0, 5, 7)
-        val listSet: MutableMultiSet<IntList> = mutableMultiSetOf(mutableList1, mutableList2)
-
-        var listIter = listSet.iterator()
-        var listExpected = listOf(listOf(1, 2, 3), listOf(0, 5, 7))
-        var listValues: MutableList<IntList> = mutableListOf()
-        while (listIter.hasNext()) {
-            listValues.add(listIter.next())
-        }
-        assertEquals(listExpected, listValues)
-
-        mutableList2.clear()
-        listIter = listSet.iterator()
-        listExpected = listOf(listOf(1, 2, 3), emptyList())
-        listValues = mutableListOf()
-        while (listIter.hasNext()) {
-            listValues.add(listIter.next())
-        }
-        assertEquals(listExpected, listValues)
+        runMultiSetMutableElementsIteratorTests(createSet())
     }
 
     @Test
